refactor(Input): rename to ts
